1. fix: resolve indexedDB not defined error during build
- Move WagmiProvider and RainbowKitProvider to separate client-only component
- Use next/dynamic with ssr: false to prevent server-side rendering
- Add loading spinner during client hydration

2. chore: suppress web-worker webpack warning

- Add ignoreWarnings config for web-worker module in next.config.ts
- Warning caused by dynamic require in ffjavascript dependency
