+++ packages/nextjs/next.config.ts
@@ -17,6 +17,10 @@ const nextConfig: NextConfig = {
   webpack: config => {
     config.resolve.fallback = { fs: false, net: false, tls: false };
     config.externals.push("pino-pretty", "lokijs", "encoding");
+
+    // Suppress web-worker dynamic require warning
+    config.ignoreWarnings = [{ module: /node_modules\/web-worker/ }];
+
     return config;
   },
 };
